mongodb-azure/iothub_to_mongodb/requirements.txt
azure-functions
pymongo
dnspython
requests